requirements.txt
fastapi
uvicorn
pydantic
python-telegram-bot